项目初始代码1.0(完成套餐模块)

+++ sky-server/src/main/java/com/sky/service/impl/SetMealServiceImpl.java
@@ -3,10 +3,15 @@ package com.sky.service.impl;
 import com.github.pagehelper.Page;
 import com.github.pagehelper.PageHelper;
 import com.google.common.collect.Lists;
+import com.sky.constant.MessageConstant;
+import com.sky.constant.StatusConstant;
 import com.sky.dto.SetmealDTO;
 import com.sky.dto.SetmealPageQueryDTO;
+import com.sky.entity.Dish;
 import com.sky.entity.Setmeal;
 import com.sky.entity.SetmealDish;
+import com.sky.exception.DeletionNotAllowedException;
+import com.sky.mapper.DishMapper;
 import com.sky.mapper.SetMealMapper;
 import com.sky.result.PageResult;
 import com.sky.service.SetMealService;
@@ -24,6 +29,8 @@ public class SetMealServiceImpl implements SetMealService {
 
     @Autowired
     SetMealMapper setMealMapper;
+    @Autowired
+    DishMapper dishMapper;
 
     /**
      * 套餐分页查询
@@ -74,6 +81,8 @@ public class SetMealServiceImpl implements SetMealService {
             //删除原有套餐
             setMealMapper.deleteByDishIds(Lists.newArrayList(setmealDTO.getId()));
 
+            setmealDTO.getSetmealDishes().stream().forEach(i -> i.setSetmealId(setmealDTO.getId()));
+
             //添加新套餐
             setMealMapper.addwithDishId(setmealDTO.getId(),setmealDTO.getSetmealDishes());
             
@@ -82,9 +91,9 @@ public class SetMealServiceImpl implements SetMealService {
         Setmeal setmeal = new Setmeal();
         BeanUtils.copyProperties(setmealDTO,setmeal);
 
+
         setMealMapper.update(setmeal);
 
-        //todo 未debug检测是否能通过 association 标签获取 setmealDishes并赋值进 SetmealVO 的属性中
         SetmealVO setmealVO = setMealMapper.queryAll(setmealDTO.getId());
 
         return setmealVO;
@@ -104,13 +113,18 @@ public class SetMealServiceImpl implements SetMealService {
         return setmealVO;
     }
 
+    /**
+     * 保存套餐
+     * @param setmealDTO
+     */
     @Override
     @Transactional
     public void save(SetmealDTO setmealDTO) {
 
-        //todo 未设置默认禁用
-        Setmeal setmeal = Setmeal.builder().build();
+        //将套餐id赋值进SetmealDishes 的 SetmealId 中
+        setmealDTO.getSetmealDishes().stream().forEach(i -> i.setSetmealId(setmealDTO.getId()));
 
+        Setmeal setmeal = Setmeal.builder().build();
         BeanUtils.copyProperties(setmealDTO,setmeal);
 
         //保存套餐及配套菜品关系
@@ -127,9 +141,30 @@ public class SetMealServiceImpl implements SetMealService {
     @Override
     public void changeStatus(Integer status, Long id) {
 
+        //更改菜品状态
         Setmeal setmeal = Setmeal.builder().status(status).id(id).build();
 
-        setMealMapper.update(setmeal);
+        //判断是否状态改为起售 获取套餐中菜品的状态，有未起售则不允许套餐起售
+        if(status.equals(StatusConstant.ENABLE)){
+
+            SetmealVO setmealVO = setMealMapper.queryAll(id);
+            List<Long> dishesIds = setmealVO.getSetmealDishes().stream()
+                    .map(SetmealDish::getDishId)
+                    .collect(Collectors.toList());
+
+            //获取状态为未起售的菜品数量
+            List<Dish> dishList = dishMapper.bulkQueriesById(dishesIds);
+            long enablesCount = dishList.stream().filter(i -> i.getStatus().equals(StatusConstant.DISABLE)).count();
+
+            if(enablesCount != 0) {
+                throw new DeletionNotAllowedException(MessageConstant.SETMEAL_ENABLE_FAILED);
+            }else{
+                setMealMapper.update(setmeal);
+            }
+
+        }else{
+            setMealMapper.update(setmeal);
+        }
 
     }
 
@@ -141,10 +176,19 @@ public class SetMealServiceImpl implements SetMealService {
     @Transactional
     public void batchDelete(List<Long> ids) {
 
-        //删除本表
-        setMealMapper.batchDeletion(ids);
+        List<Setmeal> setmeals = setMealMapper.batchQuery(ids);
+        long statusCount = setmeals.stream().filter(i -> i.getStatus().equals(StatusConstant.ENABLE)).count();
 
-        //删除配套菜品关系表
-        setMealMapper.deleteBySetmealIds(ids);
+        if(statusCount != 0){
+            throw new DeletionNotAllowedException(MessageConstant.SETMEAL_ON_SALE);
+        }else{
+
+            //删除本表
+            setMealMapper.batchDeletion(ids);
+            //删除配套菜品关系表
+            setMealMapper.deleteBySetmealIds(ids);
+
+        }
+
     }
 }
